# [language][diem-framework] refactor artifact generation