--- a/language/diem-tools/writeset-transaction-generator/src/admin_script_builder.rs
+++ b/language/diem-tools/writeset-transaction-generator/src/admin_script_builder.rs
@@ -2,19 +2,37 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::Result;
-use diem_framework::compile_script;
 use diem_types::{
     account_address::AccountAddress,
     account_config::diem_root_address,
     transaction::{Script, WriteSetPayload},
 };
 use handlebars::Handlebars;
+use move_lang::compiled_unit::CompiledUnit;
 use serde::Serialize;
 use std::{collections::HashMap, io::Write, path::PathBuf};
 use tempfile::NamedTempFile;
 
 /// The relative path to the scripts templates
 pub const SCRIPTS_DIR_PATH: &str = "templates";
+
+pub fn compile_script(source_file_str: String) -> Vec<u8> {
+    let (_files, mut compiled_program) = move_lang::move_compile_and_report(
+        &[source_file_str],
+        &diem_framework::diem_stdlib_files(),
+        Some(move_lang::shared::Address::DIEM_CORE),
+        None,
+        false,
+    )
+    .unwrap();
+    let mut script_bytes = vec![];
+    assert!(compiled_program.len() == 1);
+    match compiled_program.pop().unwrap() {
+        CompiledUnit::Module { .. } => panic!("Unexpected module when compiling script"),
+        CompiledUnit::Script { script, .. } => script.serialize(&mut script_bytes).unwrap(),
+    };
+    script_bytes
+}
 
 fn compile_admin_script(input: &str) -> Result<Script> {
     let mut temp_file = NamedTempFile::new()?;