This cleans up the code to generate various artifacts, in preparation for the incoming `diem-framework-releases` crate:
- Each pass is now a separate function, whose output path can now be specified as an argument instead of being hard-coded. This will make it easier for us to change the directory layouts and allow the tool to output to different locations in the future, including ones dynamically specified by the user through the command-line.
- Some passes run faster now
  - Script ABI Generation: 
    - 60s -> 10s (debug)
    - 6s -> 1s (release)
  - Script Compilation:
    - 25s -> 3.6s (debug)
    - 3.3s -> 0.5s (release)